Use actual ACF 5.8 API
Since this package was built with beta versions of 5.8, we were using
the old `acf_register_block` function which seems to have switched names
to `acf_register_block_type`. It was still working, but we should be
using the official function.

Also, this adds support for a few more options in the API:

- align: `->setAlign('full')`
- mode: `->setMode('preview')`
- post types: `->setPostTypes('page', 'post')`
- supports: `->setSupports([ 'mode' => 'preview', 'align' => 'full', 'multiple' => 'false' ])`

Might add the enqueue stuff in the future, but I personally think they
aren’t the best idea. I can be persuaded.